// src/shared/enums/countryEnum.ts
export enum CountryEnum {
  generic = "generic",
  mexico = "Mexico",
  ecuador = "Ecuador",
  brasil = "Brazil",
  usa = "EEUU",
  chile = "Chile",
  panama = "Panama",
  peru = "Peru",
  el_salvador = "ElSalvador",
  honduras = "Honduras",
  guatemala = "Guatemala",
  nicaragua = "Nicaragua",
  costa_rica = "CostaRica",
  colombia = "Colombia",
}

export enum CountryTitleEnum {
  mexico = "México",
  panama = "Panamá",
  brasil = "Brasil",
  peru = "Perú",
  costa_rica = "Costa Rica",
  nicaragua = "Nicaragua",
  guatemala = "Guatemala",
  el_salvador = "El Salvador",
}